+++ lib/src/styled_toast_theme.dart
@@ -17,95 +17,91 @@ class StyledToastTheme extends InheritedWidget {
   final Widget child;
 
   ///Text align
-  final TextAlign textAlign;
+  final TextAlign? textAlign;
 
   ///Text direction
-  final TextDirection textDirection;
+  final TextDirection? textDirection;
 
   ///Border radius
-  final BorderRadius borderRadius;
+  final BorderRadius? borderRadius;
 
   ///Background color
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   ///Padding for the text and the container edges
-  final EdgeInsets textPadding;
+  final EdgeInsets? textPadding;
 
   ///Text style for content
-  final TextStyle textStyle;
+  final TextStyle? textStyle;
 
   ///Shape for the container
-  final ShapeBorder shapeBorder;
+  final ShapeBorder? shapeBorder;
 
   ///Toast show duration
-  final Duration duration;
+  final Duration? duration;
 
   ///Toast animation duration
-  final Duration animDuration;
+  final Duration? animDuration;
 
   ///Position of the toast widget in current window
-  final StyledToastPosition toastPositions;
+  final StyledToastPosition? toastPositions;
 
   ///Alignment of animation, like size, rotate animation.
-  final AlignmentGeometry alignment;
+  final Alignment? alignment;
 
   ///Axis of animation, like size animation
-  final Axis axis;
+  final Axis? axis;
 
   ///Start offset of slide animation
-  final Offset startOffset;
+  final Offset? startOffset;
 
   ///End offset of slide animation
-  final Offset endOffset;
+  final Offset? endOffset;
 
   ///Start offset of reverse slide animation
-  final Offset reverseStartOffset;
+  final Offset? reverseStartOffset;
 
   ///End offset of reverse slide animation
-  final Offset reverseEndOffset;
+  final Offset? reverseEndOffset;
 
   ///Toast animation
-  final StyledToastAnimation toastAnimation;
+  final StyledToastAnimation? toastAnimation;
 
   ///Toast reverse animation
-  final StyledToastAnimation reverseAnimation;
+  final StyledToastAnimation? reverseAnimation;
 
   ///Animation curve
-  final Curve curve;
+  final Curve? curve;
 
   ///Animation reverse curve
-  final Curve reverseCurve;
+  final Curve? reverseCurve;
 
   ///Dismiss old toast when new one shows
-  final bool dismissOtherOnShow;
-
-  ///When window change, moving toast.
-  @deprecated
-  final bool movingOnWindowChange;
+  final bool? dismissOtherOnShow;
 
   ///Callback when toast dismissed
-  final VoidCallback onDismiss;
+  final VoidCallback? onDismiss;
 
   ///Full width that the width of the screen minus the width of the margin.
-  final bool fullWidth;
+  final bool? fullWidth;
 
   ///Is hide keyboard when toast show
-  final bool isHideKeyboard;
+  final bool? isHideKeyboard;
 
   ///Custom animation builder method
-  final CustomAnimationBuilder animationBuilder;
+  final CustomAnimationBuilder? animationBuilder;
 
   ///Custom animation builder method
-  final CustomAnimationBuilder reverseAnimBuilder;
+  final CustomAnimationBuilder? reverseAnimBuilder;
 
   ///Is the input ignored for the toast.
-  final bool isIgnoring;
+  final bool? isIgnoring;
 
   ///When toast widget [initState], this callback will be called.
-  final OnInitStateCallback onInitState;
+  final OnInitStateCallback? onInitState;
 
   StyledToastTheme({
-    this.child,
+    required this.child,
     this.textAlign,
     this.textDirection,
     this.borderRadius,
@@ -127,7 +123,6 @@ class StyledToastTheme extends InheritedWidget {
     this.curve,
     this.reverseCurve,
     this.dismissOtherOnShow,
-    this.movingOnWindowChange,
     this.onDismiss,
     this.fullWidth,
     this.isHideKeyboard,
@@ -143,6 +138,6 @@ class StyledToastTheme extends InheritedWidget {
     return true;
   }
 
-  static StyledToastTheme of(BuildContext context) =>
+  static StyledToastTheme? of(BuildContext context) =>
       context.dependOnInheritedWidgetOfExactType<StyledToastTheme>();
 }